refactor search component

+++ features/home_search/lib/src/domain/entities/item_entity.dart
@@ -26,8 +26,8 @@ class FollowersEntity {
 }
 
 class ImagesEntity {
-  final double heigth;
-  final double width;
+  final int heigth;
+  final int width;
   final String url;
   ImagesEntity({
     required this.heigth,
